+++ festival/src/main/java/com/example/festival/partner/repository/PartnerRepositoryInterface.java
@@ -3,5 +3,5 @@ package com.example.festival.partner.repository;
 import com.example.festival.partner.entity.Partner;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PartnerRepositoryInterface extends CrudRepository<Partner, Long> {
+public interface PartnerRepositoryInterface extends CrudRepository<Partner, Integer> {
 }
